+++ lib/rdf/rdfa.rb
@@ -28,10 +28,11 @@ module RDF
     require 'rdf/rdfa/patches/literal_hacks'
     require 'rdf/rdfa/patches/nokogiri_hacks'
     require 'rdf/rdfa/patches/string_hacks'
-    autoload :Profile,  'rdf/rdfa/profile'
-    autoload :Reader,  'rdf/rdfa/reader'
-    autoload :Writer,  'rdf/rdfa/writer'
-    autoload :VERSION, 'rdf/rdfa/version'
+    autoload :Expansion,  'rdf/rdfa/expansion'
+    autoload :Profile,    'rdf/rdfa/profile'
+    autoload :Reader,     'rdf/rdfa/reader'
+    autoload :Writer,     'rdf/rdfa/writer'
+    autoload :VERSION,    'rdf/rdfa/version'
     
     XML_RDFA_PROFILE    = "http://www.w3.org/profile/rdfa-1.1"
     XHTML_RDFA_PROFILE  = "http://www.w3.org/profile/html-rdfa-1.1"